Add Filterable concern in order to specify common methods from meeting controllers
