+++ decidim-meetings/app/controllers/decidim/meetings/directory/meetings_controller.rb
@@ -8,6 +8,7 @@ module Decidim
         layout "layouts/decidim/application"
 
         include FilterResource
+        include Filterable
         include Paginable
 
         helper Decidim::WidgetUrlsHelper
@@ -63,25 +64,8 @@ module Decidim
           %w(all) + current_organization.public_participatory_spaces.collect(&:model_name).uniq.collect(&:name).collect(&:underscore)
         end
 
-        def default_filter_type_params
-          %w(all) + Decidim::Meetings::Meeting::TYPE_OF_MEETING
-        end
-
         def default_filter_scope_params
           %w(all global) + current_organization.scopes.pluck(:id).map(&:to_s)
-        end
-
-        def default_search_params
-          {
-            scope: Meeting.not_hidden.visible_meeting_for(current_user)
-          }
-        end
-
-        def default_filter_origin_params
-          filter_origin_params = %w(citizens)
-          filter_origin_params << "official"
-          filter_origin_params << "user_group" if current_organization.user_groups_enabled?
-          filter_origin_params
         end
 
         def context_params
