+++ decidim-meetings/app/controllers/decidim/meetings/meetings_controller.rb
@@ -5,10 +5,12 @@ module Decidim
     # Exposes the meeting resource so users can view them
     class MeetingsController < Decidim::Meetings::ApplicationController
       include FilterResource
+      include Filterable
       include Flaggable
       include Withdrawable
       include FormFactory
       include Paginable
+
       helper Decidim::WidgetUrlsHelper
       helper Decidim::ResourceVersionsHelper
 
@@ -131,23 +133,6 @@ module Decidim
           state: nil,
           origin: default_filter_origin_params,
           type: default_filter_type_params
-        }
-      end
-
-      def default_filter_origin_params
-        filter_origin_params = %w(citizens)
-        filter_origin_params << "official"
-        filter_origin_params << "user_group" if current_organization.user_groups_enabled?
-        filter_origin_params
-      end
-
-      def default_filter_type_params
-        %w(all) + Decidim::Meetings::Meeting::TYPE_OF_MEETING
-      end
-
-      def default_search_params
-        {
-          scope: Meeting.not_hidden.visible_meeting_for(current_user)
         }
       end
     end
