+++ decidim-meetings/app/helpers/decidim/meetings/directory/application_helper.rb
@@ -91,8 +91,6 @@ module Decidim
 
             next if categories_values.empty?
 
-            next if categories_values.empty?
-
             key_point = current_participatory_space.class.name.gsub("::", "__") + current_participatory_space.id.to_s
 
             TreeNode.new(
